+++ src/search.h
@@ -44,16 +44,16 @@ typedef struct t_ts_file_match
 {
 	ts_found_file* file;
 	int line_nr;
-	int word_match_offset; // nr of bytes, not codepoints.
-	int word_match_length; // nr of bytes, not codepoints.
+	size_t word_match_offset; // nr of bytes, not codepoints.
+	size_t word_match_length; // nr of bytes, not codepoints.
 	utf8_int8_t *line_info;
 } ts_file_match;
 
 typedef struct t_ts_text_match
 {
 	int line_nr;
-	int word_offset;
-	int word_match_len;
+	size_t word_offset;
+	size_t word_match_len;
 	utf8_int8_t *line_start;
 	utf8_int8_t *line_info;
 } ts_text_match;
@@ -61,7 +61,7 @@ typedef struct t_ts_text_match
 extern ts_search_result* current_search_result;
 
 ts_array 			ts_get_filters(utf8_int8_t *pattern);
-int 				ts_filter_matches(ts_array *filters, utf8_int8_t *string, utf8_int8_t **matched_filter);
+size_t 				ts_filter_matches(ts_array *filters, utf8_int8_t *string, utf8_int8_t **matched_filter);
 int 				ts_string_match(utf8_int8_t *first, utf8_int8_t *second);
 ts_search_result* 	ts_create_empty_search_result();
 bool 				ts_string_contains(utf8_int8_t *text_to_search, utf8_int8_t *text_to_find, ts_array *text_matches);
